+++ Core/src/com/mfino/domain/ChargetxnTransferMap.java
@@ -2,8 +2,6 @@ package com.mfino.domain;
 
 // Generated Sep 27, 2016 5:23:21 PM by Hibernate Tools 3.4.0.CR1
 
-import java.math.BigDecimal;
-
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -23,7 +21,7 @@ public class ChargetxnTransferMap extends Base implements java.io.Serializable {
 	public static final String FieldName_SctlId = "sctlid";
 	public static final String FieldName_CommodityTransferID = "commoditytransferid";
 	
-	private BigDecimal sctlid;
+	private Long sctlid;
 	private Long commoditytransferid;
 	private Long id;
 
@@ -44,11 +42,11 @@ public class ChargetxnTransferMap extends Base implements java.io.Serializable {
 
 	
 	@Column(name = "SCTLID", nullable = false, scale = 0)
-	public BigDecimal getSctlid() {
+	public Long getSctlid() {
 		return this.sctlid;
 	}
 
-	public void setSctlid(BigDecimal sctlid) {
+	public void setSctlid(Long sctlid) {
 		this.sctlid = sctlid;
 	}
